Move Swap state out of storage
